requirements.txt
django
djangorestframework
django-cors-headers
celery
redis
pycryptodome
cryptography
python-dotenv
boto3
django-celery-beat
django-celery-results
pyjwt
bcrypt
djangorestframework-simplejwt